make selection progression more user friendly and intuitive
