+++ src/components/Header/Header.js
@@ -1,11 +1,11 @@
 import React from 'react';
 
-const Header = ({ teamSize }) => (
+const Header = () => (
     <header className="main-header">
         <h1 className="main-header-text">
             <span className="accent-color">C</span>ome <span className="accent-color">O</span>n <span className="accent-color">Y</span>ou _____ !
         </h1>
-        <p>Add players to the pitch to assign them randomly to { teamSize }-a-side football teams</p>
+        <p>Select your team size, add players to the pitch to assign them randomly to teams, then set your team name.</p>
     </header>
 )
