+++ src/components/PlayerEntry/PlayerEntry.js
@@ -45,7 +45,7 @@ const PlayerEntry = ({ handleSubmit, teamsFilled }) => {
         </form>
         : // once teams are full, game on!
         <h3 className="pitch-header">
-            Teams Complete, Game On!
+            Teams full, Game On!
         </h3>
     )
 }
