+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Provider } from "react-redux";
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './App/';
 import store from './data/store';
 
 // CSS
